Update BackgroundTask.java
Fixed the ip issue. Now (using the default username/password we have set up) you can log in to the database with the app! The big thing with this is the proof of concept - we now can successfully pass variables from the java app and read them in the php file. This includes entire strings that can be read as statements to add info to the DB tables